requirements.txt
certifi==2020.6.20
cycler==0.10.0
kiwisolver==1.2.0
numpy==1.19.2
pandas==1.1.3
Pillow==8.0.1
pkg-resources==0.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
